ml_project/requirements.txt
click==8.0.3
easydict==1.10
hydra-core==1.2.0
joblib==1.2.0
numpy==1.23.3
omegaconf==2.2.3
pandas==1.5.0
PyYAML==6.0
scikit_learn==1.1.3
sweetviz==2.1.4
